Add MSRV-version to for `1.x` release track
I've noticed, that there is no MSRV specified for `1.x`-versions of this
crate, which is unfortunate, as this prevents the MSRV-aware resolver of
Cargo 1.84+ to check, whether the crate is available or not. Therefore
this commit adds the missing meta-data.

I've selected Rust 1.61, since this is the lowest version, where
`cargo build --features=serde` is successful.